--- therapist/src/main/java/com/projectTeam/therapist/boardService/BoardService.java
@@ -1,8 +1,0 @@
-package com.projectTeam.therapist.boardService;
-
-import com.projectTeam.therapist.model.ReplyDto;
-
-
-public interface BoardService {
-    public void writeReply(ReplyDto reply);
-}
